Move reusable cell-menu helpers to gui_helpers and restore copy toast
Move the shared right-click hit-box and copy context-menu helpers added in
this PR (CellMenuTarget, PositionCell, RenderRowHitbox, CaptureCellRightClick,
BeginCellContextMenu, EndCellContextMenu, AddCopyRowCellMenuItems) out of
rocprofvis_events_view and into rocprofvis_gui_helpers so other views (e.g.
compute) can reuse them instead of re-implementing the pattern.

Also restore the copy notification: AddCopyRowCellMenuItems now raises the
COPY_DATA_NOTIFICATION toast on both "Copy Row Data" and "Copy Cell Data".
